Aperture Shape Generation Based on AdaGrad
There are two folders in this dataset.

There are four files in the code folder. First, the data of CERR simulation irradiation is saved for subsequent optimization by VS. And we use VS to optimize and provide the C code of the core formula of this study. Then the optimized results are displayed using CERR. Finally, the code for calculating NTCP and gEUD is provided.

In the optimized data folder, the optimized result files of H1, H2, P1 and P2 cases in the paper are provided, which can be used to display DVH and dose in the CERR using the read code in the code folder.

Finally, it is about the explanation of the real patient data involved in the paper, which is not easy to disclose because of the privacy.